==> 10-conclusion.Rmd <==
# Conclusion

## Takeaways

If insurance companies were to stop using height and weight as part of their models for calculating insurance rates, a shift toward utilizing behavioral and lifestyle factors such as dietary habits and physical activities would become increasingly important. This approach would enable them to effectively market insurance products to individuals more likely to require their services, thereby potentially increasing the associated costs.

With our models, we could accurately predict potential obese individuals using only lifestyle factors, achieving high accuracy levels of 90%. Insurance companies, by leveraging our modeling pipeline, could price health insurance premiums more appropriately.

By integrating the typical additional healthcare costs of $2,500 for obese individuals with the probability of obesity calculated by our model, insurers could refine their premium policy adjustments to better reflect individual risk profiles, thus potentially enhancing their financial outlook.

We would advise insurance companies to prioritize stacked models without the cost matrix to minimize false negatives, ensuring that high-risk individuals are accurately identified. This approach could lead to a reduction in overall insurance payouts.


## Business Impact - Recommendations

Here are some additional recommendations we will provide based on this analysis:

  - Improve the model by incorporating additional behavioral variables such as sleep patterns and occupational details. This could enhance risk prediction accuracy and prepare the companies for potential regulatory changes that might limit the use of traditional metrics like BMI.

  - Develop incentive programs that reward insured clients for measurable health improvements and engagement in preventive health programs. These incentives could take the form of premium reductions or rebates, aligning financial savings with healthy outcomes.

  - Adjust policies to ensure that premium adjustments are fair and transparently communicated to customers. This would provide trust in health management programs, ultimately leading to reduced insurance claims.

We believe it is important to continuously monitor the effectiveness of the model on real world data and its impact on insurance payouts.
